+++ ToDoListClass.h
@@ -13,11 +13,11 @@ public:
     void addElement(const std::string& value);
 
     // Edits an element in the to do list
-    void editElement(size_t elementId, const std::string& value);
+    void editElement(int elementId, const std::string& value);
 
     // Removes an element from the to do list
-    void removeElement(size_t elementId) {
-        if (elementId < 1 || elementId > list.size()) {
+    void removeElement(int elementId) {
+        if (elementId < 1 || elementId > static_cast<int>(list.size())) {
             std::printf("Invalid element ID.");
             return;
         }
